requirements.txt
pymongo == 4.2.0
pandas
numpy
scikit-learn
pymongo
python-dotenv

-e . #editable mode